feat: Add tag management system
- Create tag_commands.rs module with separated tag command logic
- Implement add_tags function for adding tags to projects
- Add remove_tags function for removing tags from projects
- Create list_tags function for displaying all tags with usage counts
- Add show_tags function for displaying tags of specific projects
- Support current directory project detection for tag show command
- Refactor main.rs to use modular tag command functions
