+++ src/main.rs
@@ -6,9 +6,11 @@ use uuid::Uuid;
 
 mod config;
 mod utils;
+mod tag_commands;
 
 use config::{get_config_path, load_config, save_config, Config};
 use utils::get_last_git_commit_time;
+use tag_commands::{add_tags, remove_tags, list_tags, show_tags};
 
 #[derive(Parser)]
 #[command(author, version, about, long_about = None)]
@@ -216,74 +218,21 @@ async fn main() {
         Commands::Tag { action } => match action {
             TagAction::Add { project_name, tags } => {
                 let mut config = load_config().await.unwrap();
-                if let Some(project) = config.find_project_by_name_mut(project_name) {
-                    for tag in tags {
-                        if !project.tags.contains(tag) {
-                            project.tags.push(tag.clone());
-                        }
-                    }
-                    project.updated_at = Utc::now();
-                    save_config(&config).await.unwrap();
-                    println!("✓ Tags added to project '{}'", project_name);
-                } else {
-                    println!("❌ Project not found: {}", project_name);
-                }
+                add_tags(project_name, tags, &mut config).await.unwrap();
+                save_config(&config).await.unwrap();
             }
             TagAction::Remove { project_name, tags } => {
                 let mut config = load_config().await.unwrap();
-                if let Some(project) = config.find_project_by_name_mut(project_name) {
-                    let initial_tags_count = project.tags.len();
-                    project.tags.retain(|tag| !tags.contains(tag));
-                    if project.tags.len() < initial_tags_count {
-                        project.updated_at = Utc::now();
-                        save_config(&config).await.unwrap();
-                        println!("✓ Tags removed from project '{}'", project_name);
-                    } else {
-                        println!("No matching tags found to remove from project '{}'.", project_name);
-                    }
-                } else {
-                    println!("❌ Project not found: {}", project_name);
-                }
+                remove_tags(project_name, tags, &mut config).await.unwrap();
+                save_config(&config).await.unwrap();
             }
             TagAction::Ls {} => {
                 let config = load_config().await.unwrap();
-                let mut tag_counts: std::collections::HashMap<String, u32> = std::collections::HashMap::new();
-
-                for project in config.projects.values() {
-                    for tag in &project.tags {
-                        *tag_counts.entry(tag.clone()).or_insert(0) += 1;
-                    }
-                }
-
-                let mut sorted_tags: Vec<(String, u32)> = tag_counts.into_iter().collect();
-                sorted_tags.sort_by(|a, b| b.1.cmp(&a.1));
-
-                println!("All Tags (by usage count):");
-                for (tag, count) in sorted_tags {
-                    println!("  - {} ({} projects)", tag, count);
-                }
+                list_tags(&config).await.unwrap();
             }
             TagAction::Show { project_name } => {
                 let config = load_config().await.unwrap();
-                let project_to_show = if let Some(name) = project_name {
-                    config.find_project_by_name(&name)
-                } else {
-                    let current_path = std::env::current_dir().unwrap();
-                    config.find_project_by_path(&current_path)
-                };
-
-                if let Some(project) = project_to_show {
-                    println!("Tags for '{}':", project.name);
-                    if project.tags.is_empty() {
-                        println!("  (No tags)");
-                    } else {
-                        for tag in &project.tags {
-                            println!("  - {}", tag);
-                        }
-                    }
-                } else {
-                    println!("❌ Project not found. Please specify a project name or navigate into a project directory.");
-                }
+                show_tags(project_name.as_deref(), &config).await.unwrap();
             }
         },
         Commands::Init {} => {
